Fixes in error handling

+++ src/middlewares/validatePropertyId.ts
@@ -1,5 +1,5 @@
 import { body } from 'express-validator'
 
 export const validatePropertyId = [
-  body('propertyId').isInt({ min: 1 }).toInt().withMessage('propertyId must be a positive integer'),
+  body('propertyId').isInt({ min: 1 }).withMessage('propertyId must be a positive integer'),
 ]
